2025EM1100026/ecommerce_seller_recommendation/local/pyproject.toml
[project]
name = "ecommerce-seller-recommendation"
version = "1.0.0"
description = "E-commerce Top-Seller Items Recommendation System - MSc Data Science & AI Assignment"
authors = [
    {name = "Student 2025EM1100026", email = "student@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyspark==3.5.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
